+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 - All global variables
 - Crearted function to build the navbar menu
 - Add functionality to distinguish the section in view.
+- Scroll to section when navbar item is clicked
 
 ## Changed
 
 - Connected javascript code to index.html
 - Renamed ".your-active-class" to ".active-section"
 - Created 6 sections in index.html
+- Fixed inView function (activate element when it's fully visible)
 
 ## Removed